[minor/compile] kill warning in a portable way. The proper fix (use %td in the format instead of %d / %ld) is only available from glibc 2.1 on, so can't be used portably
git-svn-id: https://subversion.renater.fr/wimsdev/trunk@3813 b48fb02d-6135-0410-b1dd-deb767c81da9

+++ wims/src/Misc/mathexp/extract.c
@@ -29,7 +29,7 @@ void extrout(char *pb, char *pe, int dist, int commas[], int *commacnt)
 	if(!checkregex(buf)) return;
     }
     if(!nocoord)
-      printf("%d %d %ld, ",thisobjline,dist+_estart,pe-pb+dist+_estart);
+      printf("%d %d %d, ",thisobjline,dist+_estart,(int)(pe-pb)+dist+_estart);
     while(*commacnt>0 && commas[*commacnt-1]>pe-pb) (*commacnt)--;
     if(*commacnt>0 && !nocomma) {
 	fwrite(pb,1,commas[0],stdout); putchar(',');
